FIO-1163: fixed an issue where submission data is deleted for components with the same key as parent layout component

+++ src/middleware/submissionHandler.js
@@ -279,7 +279,8 @@ module.exports = (router, resourceName, resourceId) => {
         if (
           data &&
           component.hasOwnProperty('persistent') &&
-          !component.persistent
+          !component.persistent &&
+          !['columns', 'fieldset', 'panel', 'table', 'tabs'].includes(component.type)
         ) {
           util.deleteProp(component.key)(data);
         }
